refactor: Mapper 분리
