+++ backend/src/main/java/reviewme/review/service/ReviewGatheredLookupService.java
@@ -1,6 +1,5 @@
 package reviewme.review.service;
 
-import jakarta.annotation.Nullable;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -9,21 +8,14 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import reviewme.question.domain.OptionItem;
 import reviewme.question.domain.Question;
 import reviewme.question.repository.QuestionRepository;
 import reviewme.review.domain.Answer;
-import reviewme.review.domain.CheckboxAnswer;
-import reviewme.review.domain.CheckboxAnswerSelectedOption;
-import reviewme.review.domain.TextAnswer;
 import reviewme.review.repository.AnswerRepository;
-import reviewme.review.service.dto.response.gathered.ReviewsGatheredByQuestionResponse;
 import reviewme.review.service.dto.response.gathered.ReviewsGatheredBySectionResponse;
-import reviewme.review.service.dto.response.gathered.SimpleQuestionResponse;
-import reviewme.review.service.dto.response.gathered.TextResponse;
-import reviewme.review.service.dto.response.gathered.VoteResponse;
 import reviewme.review.service.exception.ReviewGroupNotFoundByReviewRequestCodeException;
 import reviewme.review.service.exception.SectionNotFoundInTemplateException;
+import reviewme.review.service.mapper.ReviewGatherMapper;
 import reviewme.reviewgroup.domain.ReviewGroup;
 import reviewme.reviewgroup.repository.ReviewGroupRepository;
 import reviewme.template.repository.SectionRepository;
@@ -36,6 +28,8 @@ public class ReviewGatheredLookupService {
     private final AnswerRepository answerRepository;
     private final ReviewGroupRepository reviewGroupRepository;
     private final SectionRepository sectionRepository;
+
+    private final ReviewGatherMapper reviewGatherMapper;
 
     @Transactional(readOnly = true)
     public ReviewsGatheredBySectionResponse getReceivedReviewsBySectionId(String reviewRequestCode, long sectionId) {
@@ -55,7 +49,7 @@ public class ReviewGatheredLookupService {
                 .stream()
                 .collect(Collectors.toMap(Map.Entry::getValue, entry -> questionIdAnswers.getOrDefault(entry.getKey(), List.of())));
 
-        return new ReviewsGatheredBySectionResponse(mapToResponseBySection(questionAnswers));
+        return new ReviewsGatheredBySectionResponse(reviewGatherMapper.mapToResponseBySection(questionAnswers));
     }
 
     private void validateSectionId(long sectionId, ReviewGroup reviewGroup) {
@@ -63,51 +57,5 @@ public class ReviewGatheredLookupService {
         if (!existsByIdAndTemplateId) {
             throw new SectionNotFoundInTemplateException(sectionId, reviewGroup.getTemplateId());
         }
-    }
-
-    private List<ReviewsGatheredByQuestionResponse> mapToResponseBySection(Map<Question, List<Answer>> questionsToAnswers) {
-        return questionsToAnswers.entrySet().stream()
-                .map(entry -> mapToResponseByQuestion(entry.getKey(), entry.getValue()))
-                .toList();
-    }
-
-    private ReviewsGatheredByQuestionResponse mapToResponseByQuestion(Question question, List<Answer> answers) {
-        return new ReviewsGatheredByQuestionResponse(
-                new SimpleQuestionResponse(question.getContent(), question.getQuestionType()),
-                mapToTextResponse(question, answers),
-                mapToVoteResponse(question, answers)
-        );
-    }
-
-    @Nullable
-    List<TextResponse> mapToTextResponse(Question question, List<Answer> answers) {
-        if (question.isSelectable()) {
-            return null;
-        }
-
-        return answers.stream()
-                .map(answer -> (TextAnswer) answer)
-                .map(textAnswer -> new TextResponse(textAnswer.getContent()))
-                .toList();
-    }
-
-    @Nullable
-    List<VoteResponse> mapToVoteResponse(Question question, List<Answer> answers) {
-        if (!question.isSelectable()) {
-            return null;
-        }
-
-        Map<Long, Long> voteCountByOptionItemId = answers.stream()
-                .map(answer -> (CheckboxAnswer) answer)
-                .flatMap(checkboxAnswer -> checkboxAnswer.getSelectedOptionIds().stream())
-                .collect(Collectors.groupingBy(CheckboxAnswerSelectedOption::getSelectedOptionId,
-                        Collectors.counting()));
-
-        List<OptionItem> allOptionItem = questionRepository.findAllOptionItemsById(question.getId());
-        return allOptionItem.stream()
-                .map(optionItem -> new VoteResponse(
-                        optionItem.getContent(),
-                        voteCountByOptionItemId.getOrDefault(optionItem.getId(), 0L)))
-                .toList();
     }
 }
